+++ NEWS.md
@@ -1,5 +1,7 @@
 # actR 0.1.1
 
+* added option to pass plot options to geoChronR plotting to plot.shift()
+
 # actR 0.1.0
 
 * This marks the first release of the package! Check here for updates